+++ trikot-kword/kword-flow/src/commonMain/kotlin/com/mirego/trikot/kword/flow/FlowMultiLanguageI18N.kt
@@ -1,0 +1,34 @@
+package com.mirego.trikot.kword.flow
+
+import com.mirego.trikot.kword.I18N
+import com.mirego.trikot.kword.KWordKey
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.map
+
+class FlowMultiLanguageI18N(private val initialLanguage: String, private val languages: Map<String, I18N>) : FlowI18N {
+    private val language = MutableStateFlow(initialLanguage)
+    override val i18N = language.map { languages.getValue(it) }
+    override val currentI18N get() = language.value.let { languages.getValue(it) }
+
+    override fun reset() {
+        language.value = initialLanguage
+    }
+
+    override fun changeLanguage(code: String) {
+        language.value = code
+    }
+
+    override fun getLanguageCode() = language.value
+
+    override operator fun get(key: KWordKey) =
+        i18N.map { it[key] }
+
+    override fun t(key: KWordKey) =
+        i18N.map { it.t(key) }
+
+    override fun t(key: KWordKey, vararg arguments: Pair<String, String>) =
+        i18N.map { it.t(key, *arguments) }
+
+    override fun t(key: KWordKey, count: Int, vararg arguments: Pair<String, String>) =
+        i18N.map { it.t(key, count, *arguments) }
+}