+++ trikot-kword/kword-streams/src/commonMain/kotlin/com/mirego/trikot/kword/streams/DynamicI18N.kt
@@ -1,5 +1,7 @@
-package com.mirego.trikot.kword
+package com.mirego.trikot.kword.streams
 
+import com.mirego.trikot.kword.I18N
+import com.mirego.trikot.kword.KWordKey
 import com.mirego.trikot.streams.reactive.BehaviorSubject
 import org.reactivestreams.Publisher