+++ trikot-kword/sample/common/src/commonMain/kotlin/com/mirego/kword/sample/ServiceLocatorImpl.kt
@@ -1,0 +1,7 @@
+package com.mirego.kword.sample
+
+import com.mirego.trikot.kword.flow.FlowI18N
+
+class ServiceLocatorImpl(i18N: FlowI18N) : ServiceLocator {
+    override val textProvider = TextProvider(i18N)
+}